Hardcode some contexts as they cant be resolved and fix doc strings
