+++ src/utils/vc/contexts.js
@@ -14,6 +14,7 @@ import dockPSV1Context from './contexts/dock-ps-v1.json';
 import dockPrettyVCContext from './contexts/prettyvc.json';
 import jws2020V1Context from './contexts/jws-2020-v1.json';
 import statusList21Context from './contexts/status-list-21';
+import privateStatusList21Context from './contexts/private-status-list-21';
 
 // Lookup of following URLs will lead to loading data from the context directory, this is done as the Sr25519 keys are not
 // supported in any W3C standard and vc-js has them stored locally. This is a temporary solution.
@@ -94,5 +95,9 @@ export default new Map([
   [
     'https://w3id.org/vc/status-list/2021/v1',
     statusList21Context,
+  ],
+  [
+    'https://ld.dock.io/private-status-list-21',
+    privateStatusList21Context,
   ],
 ]);
