+++ src/utils/vc/crypto/common/DockCryptoSignature.js
@@ -113,7 +113,7 @@ export default withExtendedStaticProperties(
      * @param document - JSON-LD credential
      * @param explicitProof
      * @param signingOptions
-     * @returns {(Object|*)[]}
+     * @returns {Array} - Returns [serialized cred object, cred schema]
      */
     static convertCredential({
       document,
@@ -203,7 +203,7 @@ export default withExtendedStaticProperties(
      * Convert given JSON-LD credential to TS-anoncred's credential. This would have the signature.
      * @param document - JSON-LD credential
      * @param explicitProof
-     * @returns {(|*)[]}
+     * @returns {Array}
      */
     static convertCredentialForVerification({
       document,
@@ -243,7 +243,7 @@ export default withExtendedStaticProperties(
      * Convert given JSON-LD credential to TS-anoncred's credential. This would not have the signature.
      * @param document - JSON-LD credential
      * @param explicitProof
-     * @returns {Promise<[object, CredentialSchema]>}
+     * @returns {Promise<Array>}
      */
     static async convertCredentialToSerializedForSigning({
       document,
@@ -288,6 +288,8 @@ export default withExtendedStaticProperties(
       const serializedCred = credBuilder.serializeForSigning();
 
       credBuilder.schema = CredentialSchema.generateAppropriateSchema(serializedCred, credSchema);
+
+      // Update `document` so that generated credential has `credentialSchema` and `cryptoVersion` set
       const updatedSchemaJson = credBuilder.schema.toJSON();
       serializedCred.credentialSchema = stringify(updatedSchemaJson);
       Object.assign(document.credentialSchema, updatedSchemaJson);
@@ -314,7 +316,7 @@ export default withExtendedStaticProperties(
      * Remove actual value of proof (signature) from the object and return the trimmed object and the proof value
      * @param document
      * @param explicitProof
-     * @returns {({[p: string]: *}|*)[]}
+     * @returns {Array}
      */
     static getTrimmedProofAndValue(document, explicitProof) {
       const proof = explicitProof || document.proof;
@@ -337,6 +339,12 @@ export default withExtendedStaticProperties(
       return [trimmedProof, proofVal];
     }
 
+    /**
+     * Extract schema from the document and load the schema from its id/reference if needed.
+     * @param document
+     * @param documentLoader
+     * @returns {Promise<Array>}
+     */
     static async extractSchemaForSigning(document, documentLoader) {
       let credSchema;
       // Should be false for legacy cases or when the schema is generated by CredentialBuilder
@@ -346,7 +354,7 @@ export default withExtendedStaticProperties(
         /**
          * Fetch a schema given a schema id. Currenly only fetching it from Dock blockchain
          * @param schemaId
-         * @returns {Promise<*>}
+         * @returns {Promise<Object>}
          */
         // eslint-disable-next-line no-inner-declarations
         async function getSchema(schemaId) {
